+++ src/com/intellij/idea/plugin/hybris/system/type/codeInsight/lookup/TSLookupElementFactory.kt
@@ -40,6 +40,7 @@ object TSLookupElementFactory {
     const val GROUP_2 = 2
     const val GROUP_3 = 3
     const val GROUP_4 = 4
+    const val GROUP_5 = 5
 
     fun build(
         meta: TSGlobalMetaItem,
@@ -190,6 +191,17 @@ object TSLookupElementFactory {
                 .withIcon(HybrisIcons.TypeSystem.Types.OBJECT)
         }
         ?.let { PrioritizedLookupElement.withGrouping(it, GROUP_4) }
+
+    fun buildSpecial(dom: TypeMapping?) = dom
+        ?.let {
+            val type = it.type.stringValue ?: return@let null
+            val persistenceType = it.persistenceType.stringValue ?: return@let null
+
+            LookupElementBuilder.create(type)
+                .withTypeText(persistenceType, true)
+                .withIcon(HybrisIcons.TypeSystem.Types.SPECIAL)
+        }
+        ?.let { PrioritizedLookupElement.withGrouping(it, GROUP_5) }
 
     fun buildCustomProperty(lookupString: String) = LookupElementBuilder.create(lookupString)
         .withIcon(HybrisIcons.TypeSystem.CUSTOM_PROPERTY)